Reference Source  - Added few test cases and exception handling code

+++ applications/crossreference-api/src/main/java/com/virginvoyages/crossreference/sources/ReferenceSourcesController.java
@@ -14,6 +14,11 @@ import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 import com.virginvoyages.assembly.ReferenceSourcesAssembly;
+import com.virginvoyages.crossreference.exceptions.DataInsertionException;
+import com.virginvoyages.crossreference.exceptions.DataNotFoundException;
+import com.virginvoyages.crossreference.exceptions.DataUpdationException;
+import com.virginvoyages.crossreference.exceptions.MandatoryFieldsMissingException;
+
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
 import io.swagger.annotations.ApiParam;
@@ -21,6 +26,7 @@ import io.swagger.annotations.ApiResponse;
 import io.swagger.annotations.ApiResponses;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import net.logstash.logback.encoder.org.apache.commons.lang.StringUtils;
 
 /**
  * Controller class to handle API requests for reference source methods.
@@ -57,7 +63,14 @@ public class ReferenceSourcesController {
 			@ApiParam(value = "Application identifier of client.") @RequestHeader(value = "X-VV-Client-ID", required = false) String xVVClientID) {
 		
 		log.debug("Adding Reference Source");
+		if(StringUtils.isBlank(body.referenceSource())) {
+            throw new MandatoryFieldsMissingException();
+		}
 		ReferenceSource referenceSource  = referenceSourcesAssembly.addReferenceSource(body);
+		if(null == referenceSource) {
+			log.error("ReferenceSource Not saved due to unknown reasons ==> "+body.referenceSource());
+			throw new DataInsertionException("ReferenceSource Not Saved due to Unkown reasons");
+		}
 		return new ResponseEntity<ReferenceSource>(referenceSource,HttpStatus.OK);
 	}
 
@@ -77,8 +90,12 @@ public class ReferenceSourcesController {
 			@ApiParam(value = "Correlation ID across the enterprise application components.") @RequestHeader(value = "X-Correlation-ID", required = false) String xCorrelationID,
 			@ApiParam(value = "Application identifier of client.") @RequestHeader(value = "X-VV-Client-ID", required = false) String xVVClientID) {
 		
+		if(StringUtils.isEmpty(referenceSourceID)) {
+			throw new MandatoryFieldsMissingException();
+		}
 		referenceSourcesAssembly.deleteReferenceSourceByID(referenceSourceID);
 		return new ResponseEntity<Void>(HttpStatus.OK);
+	
 	}
 
 	/**
@@ -119,7 +136,14 @@ public class ReferenceSourcesController {
 			@ApiParam(value = "Correlation ID across the enterprise application components.") @RequestHeader(value = "X-Correlation-ID", required = false) String xCorrelationID,
 			@ApiParam(value = "Application identifier of client.") @RequestHeader(value = "X-VV-Client-ID", required = false) String xVVClientID) {
 		
-		return new ResponseEntity<ReferenceSource>(referenceSourcesAssembly.findReferenceSourceByID(referenceSourceID),HttpStatus.OK);
+		if(StringUtils.isBlank(referenceSourceID)) {
+			throw new MandatoryFieldsMissingException();
+		}
+		ReferenceSource referenceSource = referenceSourcesAssembly.findReferenceSourceByID(referenceSourceID);
+		if(null == referenceSource) {
+			throw new DataNotFoundException();
+		}
+		return new ResponseEntity<ReferenceSource>(referenceSource,HttpStatus.OK);
 	}
 
 	/**
@@ -139,8 +163,15 @@ public class ReferenceSourcesController {
 			@ApiParam(value = "Correlation ID across the enterprise application components.") @RequestHeader(value = "X-Correlation-ID", required = false) String xCorrelationID,
 			@ApiParam(value = "Application identifier of client.") @RequestHeader(value = "X-VV-Client-ID", required = false) String xVVClientID) {
 		
-		
+		if(StringUtils.isBlank(body.referenceSourceID())||StringUtils.isBlank(body.referenceSource())) {
+			throw new MandatoryFieldsMissingException();
+			
+		}
 		ReferenceSource referenceSource = referenceSourcesAssembly.updateReferenceSource(body);
+		if(null == referenceSource) {
+			log.error("ReferenceSource Not saved due to unknown reasons ==> "+body.referenceSourceID());
+			throw new DataUpdationException();
+		}
 		return new ResponseEntity<ReferenceSource>(referenceSource,HttpStatus.OK);
 	}
